requirements.txt
openrouter
supabase
python-dotenv
tqdm
streamlit
requests